+++ src/groups.h
@@ -141,7 +141,8 @@ class Group {
   bool hasBLER() const;
   bool hasTime() const;
   std::chrono::time_point<std::chrono::system_clock> getRxTime() const;
-  void printHex(std::ostream* stream,
+  void printHex(std::ostream* stream) const;
+  void printHexWithTime(std::ostream* stream,
                 const std::string& time_format) const;
 
   void disableOffsets();